+++ source/Messaging.IntegrationTests/AssertOutgoingMessage.cs
@@ -15,37 +15,33 @@
 using System;
 using Dapper;
 using Messaging.Application.Configuration.DataAccess;
-using Newtonsoft.Json.Linq;
-using NodaTime;
 using Xunit;
 
 namespace Messaging.IntegrationTests
 {
     public class AssertOutgoingMessage
     {
-        private readonly IDbConnectionFactory _connectionFactory;
         private readonly dynamic _message;
 
-        private AssertOutgoingMessage(dynamic message, IDbConnectionFactory connectionFactory)
+        private AssertOutgoingMessage(dynamic message)
         {
             if (message == null) throw new ArgumentNullException(nameof(message));
             Assert.NotNull(message);
             _message = message;
-            _connectionFactory = connectionFactory;
         }
 
         public static AssertOutgoingMessage OutgoingMessage(string transactionId, string documentType, string processType, IDbConnectionFactory connectionFactory)
         {
             if (connectionFactory == null) throw new ArgumentNullException(nameof(connectionFactory));
 
-            var messages = connectionFactory.GetOpenConnection().QuerySingle(
+            var message = connectionFactory.GetOpenConnection().QuerySingle(
                 $"SELECT m.Id, m.RecordId, m.DocumentType, m.RecipientId, m.CorrelationId, m.OriginalMessageId, m.ProcessType," +
                 $"m.ReceiverRole, m.SenderId, m.SenderRole, m.ReasonCode, m.MarketActivityRecordPayload " +
                 $" FROM [b2b].[OutgoingMessages] m" +
                 $" WHERE m.OriginalMessageId = '{transactionId}' AND m.DocumentType = '{documentType}' AND m.ProcessType = '{processType}'");
 
-            Assert.NotNull(messages);
-            return new AssertOutgoingMessage(messages, connectionFactory);
+            Assert.NotNull(message);
+            return new AssertOutgoingMessage(message);
         }
 
         public AssertOutgoingMessage HasReceiverId(string receiverId)
@@ -81,41 +77,6 @@ namespace Messaging.IntegrationTests
         public AssertMarketActivityRecord WithMarketActivityRecord()
         {
             return new AssertMarketActivityRecord(_message.MarketActivityRecordPayload);
-        }
-    }
-
-    #pragma warning disable
-    public class AssertMarketActivityRecord
-    {
-        private readonly JToken _payload;
-
-        public AssertMarketActivityRecord(string payload)
-        {
-            _payload = JToken.Parse(payload);
-        }
-
-        public AssertMarketActivityRecord HasId()
-        {
-            Assert.NotNull(_payload.Value<string>("Id"));
-            return this;
-        }
-
-        public AssertMarketActivityRecord HasOriginalTransactionId(string originalTransactionId)
-        {
-            Assert.Equal(originalTransactionId, _payload.Value<string>("OriginalTransactionId"));
-            return this;
-        }
-
-        public AssertMarketActivityRecord HasMarketEvaluationPointId(string marketEvaluationPointId)
-        {
-            Assert.Equal(marketEvaluationPointId, _payload.Value<string>("MarketEvaluationPointId"));
-            return this;
-        }
-
-        public AssertMarketActivityRecord HasValidityStart(DateTime validityStart)
-        {
-            Assert.Equal(validityStart, _payload.Value<DateTime>("ValidityStart"));
-            return this;
         }
     }
 }
